Reorganized modules. Created src/cell.py to bypass circular module import error
