+++ src/screen.py
@@ -1,132 +1,145 @@
 from tkinter import (
     Tk,
-    BOTH,
-    Canvas,
-    TclError,
     Frame,
     Button,
+    BOTH,
     Entry,
-    X,
-    Y,
-    NE,
-    NS,
-    BOTTOM,
-    LEFT,
-    RIGHT,
-    TOP,
     Label,
     IntVar,
+    Canvas,
+    DISABLED,
+    NORMAL,
+    TclError,
 )
 from tkinter.messagebox import showerror
+from enum import Enum
+from typing import Tuple, Callable
 
 
-from src.utils import initialize_maze, solve_maze, render_maze
+from src.maze import Maze, MazeDrawer, MazeSolver
 
 
-class Point:
-    """Data Class that represents position on x,y grid"""
-
-    def __init__(self, x: int, y: int):
-        self._x = x
-        self._y = y
-
-    def __repr__(self):
-        return f"Point({self._x}, {self._y})"
-
-    @property
-    def x(self) -> int:
-        return self._x
-
-    @property
-    def y(self) -> int:
-        return self._y
+class State(Enum):
+    IDLE = 1
+    DRAWING = 2
+    SOLVING = 3
 
 
-class Line:
-    """
-    Class that represents distance between two points
-
-    Attributes
-    -----
-    - point1 : Point
-    - point2 : Point
-    """
-
-    def __init__(self, point1: Point, point2: Point):
-        self.__point1 = point1
-        self.__point2 = point2
-
-    def __repr__(self):
-        return f"Line(\n\t{repr(self.__point1)},\n\t{repr(self.__point2)}\n\t)"
-
-    def get_points(self) -> tuple[Point, Point]:
-        """Returns the two connecting points of a line"""
-        return self.__point1, self.__point2
-
-
-class Window:
+class Window(Frame):
     """
     Class that containig data of Tkinter window
 
     Attributes
-    -----
-    - width : int
-    - height : int
+    ----------
+    - root : Tk
+        The Tkinter root instance.
 
     Methods
-    -----
-    - wait_for_close -> None : calls self.redraw() if window still is valid
-    - start() -> None : starts mainloop for window to stay open
-    - close -> None : terminates window
-    - redraw() -> None : Updates window
-    - is_valid_window -> bool : checks if window still exists
+    -------
+    - set_state(state: State) -> None:
+        Sets the state of the window.
+    - toggle_button_state(value: str, state: Optional[bool] = None) -> None:
+        Toggles the state of a button based on the current state of the canvas.
+        If `state` is provided, sets the button state accordingly.
+    - _create_widgets() -> None:
+        Defines the frame for widget components.
+    - _create_buttons() -> None:
+        Creates buttons for drawing, solving, and resetting the maze.
+    - start() -> None:
+        Starts the Tkinter window main loop.
+    - close() -> None:
+        Terminates the window.
+    - is_valid_window() -> bool:
+        Checks if the window is still open.
+    - wait_for_close() -> None:
+        Waits for the window to close before redrawing.
     """
 
-    def __init__(self, width: int, height: int):
-        self._width = width
-        self._height = height
-        self.__root = Tk()
-        self.__root.geometry(f"{self._width}x{self._height}")
-
-        self.__canvas = CanvasFrame(self.__root)
-
-        self.__mainframe = MainFrame(self.__root, self.__canvas)
-
-        # Define frames
-        # control_frame = Frame(self.__root)
-        # control_frame.grid(row=0, column=0, sticky="ew", padx=5, pady=10)
-
-        # canvas_frame = Frame(self.__root)
-        # canvas_frame.grid(row=1, column=0, sticky="nsew")
-
-        # make canvas frame expandable
-        self.__root.grid_rowconfigure(1, weight=1)
-        self.__root.grid_columnconfigure(0, weight=1)
-
-        # initialize widgets with frames
-        # self.__btn_widgets = ButtonWidgets(control_frame)
-        # self.__entry_widgets = EntryWidgets(control_frame)
-        # self.__canvas = CanvasFrame(canvas_frame)
-
+    def __init__(self, master: Tk):
+        super().__init__(master)
+        self.__root = master
+        self.__root.geometry("800x800")
         self.__root.title("Maze Solver")
         self.__root.protocol("WM_DELETE_WINDOW", self.close)
-        # self.__canvas.pack(fill=BOTH, expand=True)
+
+        self.buttons = {}
+        self.state = State.IDLE
+
+        self._create_widgets()
+        self._create_buttons()
 
     @property
-    def root(self) -> Tk:
+    def root(self):
         return self.__root
 
-    @property
-    def width(self) -> int:
-        return self._width
+    def set_state(self, state: State):
+        self.state = state
 
-    @property
-    def height(self) -> int:
-        return self._height
+    def toggle_button_state(self, value: str, state: bool = None):
+        """
+        Checks if canvas is idle before toggling active state of buttons
+        """
+        btn = self.buttons.get(value)
+        if btn is None:
+            raise ValueError("Invalid button")
+        # if state is explicitly defined, set button to it
+        if state is not None:
+            btn["state"] = NORMAL if state else DISABLED
+        else:
+            # Only togglable when canvas is idle
+            if self.state == State.IDLE:
+                btn["state"] = NORMAL if btn["state"] == DISABLED else DISABLED
 
-    @property
-    def canvas(self) -> "CanvasFrame":
-        return self.__canvas
+    def _create_widgets(self):
+        """
+        Defines new Frame for widget components
+        """
+        self.control_frame = Frame(self.__root)
+        self.control_frame.pack()
+
+        row_label = Label(self.control_frame, text="Rows")
+        row_label.grid(row=0, column=0)
+        self.row_input = Entry(self.control_frame)
+        self.row_input.grid(row=0, column=1)
+        self.row_input.focus()
+
+        col_label = Label(self.control_frame, text="Columns")
+        col_label.grid(row=1, column=0)
+        self.col_input = Entry(self.control_frame)
+        self.col_input.grid(row=1, column=1)
+
+        self.canvas_frame = CanvasFrame(self)
+        self.canvas_frame.pack(fill=BOTH, expand=True)
+
+    def _create_buttons(self):
+        actions = ["draw", "solve", "reset"]
+        for i, action in enumerate(actions):
+            self.buttons[action] = Button(
+                self.control_frame,
+                text=f"{action.capitalize()} Maze",
+                command=getattr(self.canvas_frame, f"{action}_maze"),
+                pady=5,
+            )
+
+            self.buttons[action].grid(row=2, column=i)
+
+        if self.buttons.get("solve"):
+            self.buttons["solve"]["state"] = DISABLED
+
+    def start(self) -> None:
+        """Starts the Tkinter window"""
+        self.__root.mainloop()
+
+    def close(self) -> None:
+        """Method to terminate window"""
+        self.__root.destroy()
+
+    def is_valid_window(self) -> bool:
+        """Method that checks if window is still open before drawing to it"""
+        try:
+            return self.__root.winfo_exists()
+        except TclError:
+            pass
 
     def wait_for_close(self) -> None:
         """Method that checks if window is still open before drawing to it"""
@@ -136,184 +149,172 @@ class Window:
         except TclError:
             pass
 
-    def start(self) -> None:
-        """Starts the Tkinter window"""
-        self.__root.mainloop()
-
-    def close(self) -> None:
-        """Method to terminate window"""
-        if self.is_valid_window():
-            self.__root.destroy()
-
-    def redraw(self) -> None:
-        """
-        Method that updates Tkinter root to draw to it
-        """
-        if self.is_valid_window():
-            self.__root.update_idletasks()
-            self.__root.update()
-
-    def is_valid_window(self) -> bool:
-        """Method that checks if window is still open before drawing to it"""
-        try:
-            return self.__root.winfo_exists()
-        except TclError:
-            pass
-
-
-class MainFrame(Frame):
-    """
-    Data class that holds both CanvasFrame and Widgets. By sharing a parent, each Frame will manage their own
-    layout without conflicts
-
-    Attributes
-    -----
-    - widgets : EntryWidget
-    - canvas : CanvasFrame
-    - buttons : ButtonWidgets
-    """
-
-    def __init__(self, root: Tk, canvas_frame: "CanvasFrame"):
-        super().__init__(root)
-        self._canvas = canvas_frame
-        # self.canvas = CanvasFrame(self)
-        self.widgets = EntryWidgets(self._canvas)
-
-        # self.canvas.pack(fill=BOTH, expand=True)
-        self.widgets.pack(fill=X)
-
 
 class CanvasFrame(Frame):
     """
-    Behvaior class that handles interactions between maze and canvas
+    Behavior class that handles interactions between maze and canvas
+
+    Attributes
+    ----------
+    - window : Window
+        The parent window instance.
+
+    - canvas : Canvas
+        The Tkinter canvas instance.
+
+    - is_drawing : bool
+        Flag indicating whether drawing is in progress.
+
+    - maze : Maze
+        The maze object associated with the canvas.
 
     Methods
-    -----
-    - draw_line(line : Line, fille_color ?: str) -> None : Draws line to canvas
+    -------
+    - draw_line(line, fill_color="black") -> None:
+        Draws a line on the canvas.
+    - clear_canvas() -> None:
+        Clears the canvas.
+    - calculate_window_sizes() -> Tuple[int, int, int, int, int, int]:
+        Calculates window sizes based on user input.
+    - draw_maze(event=None) -> None:
+        Draws the maze based on user input.
+    - solve_maze(event=None) -> None:
+        Solves the maze.
+    - reset_maze() -> None:
+        Resets the maze.
+    - redraw() -> None:
+        Updates the Tkinter root to draw to it.
+    - _bind_return(func: Callable) -> None:
+        Binds the return key to a function.
     """
 
-    def __init__(self, window: Tk):
-        super().__init__(window)
+    def __init__(self, window: Window):
+        super().__init__(window.root)
         self.__window = window
-        self.__canvas = Canvas(
-            self,
-            bg="white",
-        )
-        self.__canvas.pack(fill=BOTH, expand=True)
+        self.__root = self.__window.root
+        self._bind_return(self.draw_maze)
+        self.is_drawing = False
+        self.maze = None
+        self.canvas = None
+        self.create_canvas()
 
-    def draw_line(self, line: Line, fill_color: str = "black") -> None:
+    def create_canvas(self):
+        self.canvas = Canvas(self, bg="white")
+        self.canvas.pack(fill=BOTH, expand=True)
+
+    def draw_line(self, line, fill_color="black"):
+        """Draws line to canvas"""
+        if self.__window.state in [State.DRAWING, State.SOLVING]:
+            point1, point2 = line.get_points()
+            x1, y1 = point1.x, point1.y
+            x2, y2 = point2.x, point2.y
+            self.canvas.create_line(x1, y1, x2, y2, fill=fill_color, width=2)
+
+    def clear_canvas(self):
+        self.__window.set_state(State.IDLE)
+        self.canvas.delete("all")
+
+    def calculate_window_sizes(self) -> Tuple[int, int, int, int, int, int]:
         """
-        Responsible for drawing a line on the canvas
+        Takes user input of columns and rows to calculate the size of a window based on those values.
+        Calls draw_maze with input values to draw the maze
         """
-        point1, point2 = line.get_points()
-        x1, y1 = point1.x, point1.y
-        x2, y2 = point2.x, point2.y
-        self.__canvas.create_line(x1, y1, x2, y2, fill=fill_color, width=2)
+        self.rows_entry = int(self.__window.row_input.get())
+        self.cols_entry = int(self.__window.col_input.get())
+        if (
+            self.cols_entry < 0
+            or self.cols_entry > 50
+            or self.rows_entry < 0
+            or self.rows_entry > 50
+        ):
+            raise ValueError("Maze must have between 2 and 50 columns")
+        print(f"Generating a {self.cols_entry} by {self.rows_entry} maze...")
+        desired_padding = 10
+        cell_cols = 20 if self.cols_entry < 25 else 10
+        cell_rows = 20 if self.rows_entry < 25 else 10
+        window_size = 800
+
+        maze_width = self.cols_entry * cell_cols
+        maze_height = self.rows_entry * cell_rows
+
+        window_width = maze_width + (
+            desired_padding * 2
+        )  # Maze size plus padding for left and right
+        window_height = maze_height + (
+            desired_padding * 2
+        )  # Maze size plus padding for top and bottom
+
+        padding_x = (window_size - maze_width) / 2
+        padding_y = (window_size - maze_height) / 2
+
+        return (
+            window_size,
+            window_size,
+            padding_x,
+            padding_y,
+            self.cols_entry,
+            self.rows_entry,
+        )
+
+    def draw_maze(self, event=None):
+        try:
+            width, height, padding_x, padding_y, num_cols, num_rows = (
+                self.calculate_window_sizes()
+            )
+            self.reset_maze()
+            if self.__window.buttons["solve"]["state"] == NORMAL:
+                self.__window.toggle_button_state("solve")
+            self.__window.set_state(State.DRAWING)
+
+            cell_cols = 20 if num_cols < 25 else 10
+            cell_rows = 20 if num_rows < 25 else 10
+
+            self.maze = Maze(
+                padding_x,
+                padding_y,
+                num_cols=num_cols,
+                num_rows=num_rows,
+                cell_width=cell_cols,
+                cell_height=cell_rows,
+            )
+            self.drawer = MazeDrawer(self.maze, self)
+            # enable solve button
+            self.__window.set_state(State.IDLE)
+            self.__window.toggle_button_state("solve")
+            self._bind_return(self.solve_maze)
+
+        except ValueError as e:
+            showerror("Error", message=e)
+
+    def solve_maze(self, event=None):
+        if self.__window.state == State.IDLE:
+            self.__window.set_state(State.SOLVING)
+            if self.drawer and self.maze:
+                self.__window.toggle_button_state("draw")
+                solver = MazeSolver(self.maze, self.drawer)
+                solver.solve()
+                self._bind_return(self.draw_maze)
+            else:
+                showerror(title="Error", message="Must draw maze before solving it")
+            self.__window.set_state(State.IDLE)
+            self.__window.toggle_button_state("draw")
+
+    def reset_maze(self):
+        self.clear_canvas()
+        self.__window.set_state(State.IDLE)
+        self.__window.toggle_button_state("solve", False)
+        self.__window.toggle_button_state("draw", True)
 
     def redraw(self) -> None:
         """
         Method that updates Tkinter root to draw to it
         """
-        if self.is_valid_window():
-            self.__window.update_idletasks()
-            self.__window.update()
+        if self.__window.is_valid_window():
+            self.__root.update_idletasks()
+            self.__root.update()
 
-    def is_valid_window(self) -> bool:
-        """Method that checks if window is still open before drawing to it"""
-        try:
-            return self.__window.winfo_exists()
-        except TclError:
-            pass
-
-
-class ButtonWidgets:
-    """Class that contains all buttons in the maze solver"""
-
-    def __init__(self, parent: Frame, canvas_frame: "CanvasFrame"):
-        self._container = None
-        self._parent = parent
-        self._canvas = canvas_frame
-        # Grid placement for buttons
-        self.reset_button = Button(
-            parent, text="Reset", command=self.reset, cursor="exchange"
-        )
-        self.reset_button.grid(row=1, column=0, padx=5, sticky="nsew")
-
-        self.solve_button = Button(
-            parent, text="Solve", command=self.solve, cursor="arrow"
-        )
-        self.solve_button.grid(row=2, column=0, padx=5, sticky="nsew")
-
-    def solve(self):
-        """Method that solves the maze"""
-        print(self._container)
-
-    def reset(self):
-        """Method that resets the maze"""
-        print("Resetting...")
-
-
-class EntryWidgets:
-    def __init__(self, parent):
-        self.parent = parent
-        # User input
-        self.draw_button = Button(
-            parent, text="Draw", command=self.draw, cursor="arrow"
-        )
-        self.draw_button.grid(
-            row=2, column=1, columnspan=2, ipadx=5, ipady=5, sticky="ew"
-        )
-
-        # Add entries
-        self.row_input = IntVar(value=10)
-        row_label = Label(parent, text="Rows", font=("Helvetica", 14))
-        row_label.grid(row=0, column=1, sticky="ew")
-        self.rows_entry = Entry(
-            parent,
-            textvariable=self.row_input,
-            width=5,
-            bg="white",
-            fg="black",
-            takefocus=True,
-            font=("Helvetica", 14),
-            justify="center",
-        )
-        self.rows_entry.grid(row=1, column=1, ipadx=5, ipady=5, sticky="ew")
-
-        self.col_input = IntVar(value=10)
-        col_label = Label(parent, text="Columns", font=("Helvetica", 14))
-        col_label.grid(row=0, column=2, sticky="ew")
-        self.columns_entry = Entry(
-            parent,
-            textvariable=self.col_input,
-            width=5,
-            bg="white",
-            fg="black",
-            font=("Helvetica", 14),
-            justify="center",
-        )
-        self.columns_entry.grid(row=1, column=2, ipadx=5, ipady=5, sticky="ew")
-
-        for i in range(3):
-            parent.grid_columnconfigure(i, weight=1)
-        parent.grid_rowconfigure(1, weight=2)
-
-    def draw(self):
-        """Method that draws the maze"""
-        try:
-            cols = int(self.col_input.get())
-            rows = int(self.row_input.get())
-            container = initialize_maze(cols, rows)
-            print(container)
-            self._container = container
-            render_maze(container, self.parent)
-            # self.parent.update_idletasks()
-            # self.parent.update()
-
-        except ValueError as e:
-            showerror(title="Error", message=e)
-
-
-# if __name__ == "__main__":
-#     window = Window(1000, 600)
-#     window.start()
+    def _bind_return(self, func: Callable):
+        """
+        Method that binds return key to a function
+        """
+        self.__root.bind("<Return>", func)
